feat: move interfaces
