+++ src/Dfe.PlanTech.Domain/Submissions/Interfaces/ICalculateMaturityCommand.cs
@@ -1,7 +1,6 @@
-﻿namespace Dfe.PlanTech.Application.Submissions.Interfaces
-{
+﻿namespace Dfe.PlanTech.Domain.Submissions.Interfaces;
+
     public interface ICalculateMaturityCommand
     {
         Task<int> CalculateMaturityAsync(int submissionId, CancellationToken cancellationToken = default);
     }
-}
