+++ src/Dfe.PlanTech.Domain/Users/Interfaces/ICreateEstablishmentCommand.cs
@@ -1,6 +1,6 @@
 using Dfe.PlanTech.Domain.Establishments.Models;
 
-namespace Dfe.PlanTech.Application.Persistence.Interfaces;
+namespace Dfe.PlanTech.Domain.Users.Interfaces;
 
 public interface ICreateEstablishmentCommand
 {
