+++ src/Dfe.PlanTech.Domain/Users/Interfaces/IUser.cs
@@ -1,0 +1,12 @@
+﻿using Dfe.PlanTech.Domain.Establishments.Models;
+
+namespace Dfe.PlanTech.Domain.Users.Interfaces;
+
+public interface IUser
+{
+    Task<int?> GetCurrentUserId();
+
+    Task<int> GetEstablishmentId();
+
+    EstablishmentDto GetOrganisationData();
+}
